Fixed confine bug when digest occurs
When in "confine" mode, set upper limit to the top of parent element.

There was a bug when our sticky element has been moved to the bottom (by
scrolling) and digest occurs, the new upper limit became the top of the
sticky element, which is only the sticky elements height away from the
stickyBottomLine.